# docs/openapi.yaml
openapi: 3.1.0
info:
  title: x4-api
  version: 0.0.1
paths:
...
  /api/v1/modules:
    get:
      tags:
      - modules
      summary: List Modules
      operationId: list_modules_api_v1_modules_get
      parameters:
      - name: kind
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: 'Filter by kind: production, habitation, storage, dock, defence,
            connectionmodule'
          title: Kind
        description: 'Filter by kind: production, habitation, storage, dock, defence,
          connectionmodule'
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 2000
          minimum: 1
          default: 500
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ModuleSummary'
                title: Response List Modules Api V1 Modules Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/modules/{module_id}:
    get:
      tags:
      - modules
      summary: Get Module
      operationId: get_module_api_v1_modules__module_id__get
      parameters:
      - name: module_id
        in: path
        required: true
        schema:
          type: string
          title: Module Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModuleDetail'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/ships:
    get:
      tags:
      - ships
      summary: List Ships
      description: List all ships in the game catalog.
      operationId: list_ships_api_v1_ships_get
      parameters:
      - name: class_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Class Id
      - name: faction_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Faction Id
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 2000
          minimum: 1
          default: 500
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ShipSummary'
                title: Response List Ships Api V1 Ships Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/ships/{ship_id}:
    get:
      tags:
      - ships
      summary: Get Ship
      description: Get detailed stats for a specific ship.
      operationId: get_ship_api_v1_ships__ship_id__get
      parameters:
      - name: ship_id
        in: path
        required: true
        schema:
          type: string
          title: Ship Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShipDetail'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/npc-stations:
    get:
      tags:
      - npc-stations
      summary: List Npc Stations
      description: List NPC stations in the universe, optionally filtered by owner,
        tag, or sector.
      operationId: list_npc_stations_api_v1_npc_stations_get
      parameters:
      - name: owner
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Filter by owner faction ID
          title: Owner
        description: Filter by owner faction ID
      - name: tag
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Filter by tag, e.g. shipyard, wharf, equipmentdock
          title: Tag
        description: Filter by tag, e.g. shipyard, wharf, equipmentdock
      - name: sector
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Filter by location_sector (case-insensitive)
          title: Sector
        description: Filter by location_sector (case-insensitive)
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/NpcStationSummary'
                title: Response List Npc Stations Api V1 Npc Stations Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/npc-stations/{station_id}:
    get:
      tags:
      - npc-stations
      summary: Get Npc Station
      description: Get full details for a specific NPC station including coordinates.
      operationId: get_npc_station_api_v1_npc_stations__station_id__get
      parameters:
      - name: station_id
        in: path
        required: true
        schema:
          type: string
          title: Station Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NpcStationDetail'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ModuleDetail:
      properties:
        module_id:
          type: string
          title: Module Id
        name:
          type: string
          title: Name
        dlc:
          anyOf:
          - type: string
          - type: 'null'
          title: Dlc
        kind:
          anyOf:
          - type: string
          - type: 'null'
          title: Kind
        size:
          anyOf:
          - type: string
          - type: 'null'
          title: Size
        produces_ware_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Produces Ware Id
        storage_capacity:
          anyOf:
          - type: integer
          - type: 'null'
          title: Storage Capacity
        storage_type:
          anyOf:
          - type: string
          - type: 'null'
          title: Storage Type
        drone_capacity:
          anyOf:
          - type: integer
          - type: 'null'
          title: Drone Capacity
        workforce_capacity:
          anyOf:
          - type: integer
          - type: 'null'
          title: Workforce Capacity
        icon_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Icon Url
        hull:
          anyOf:
          - type: integer
          - type: 'null'
          title: Hull
        explosiondamage:
          anyOf:
          - type: integer
          - type: 'null'
          title: Explosiondamage
        turrets_s:
          type: integer
          title: Turrets S
        turrets_m:
          type: integer
          title: Turrets M
        turrets_l:
          type: integer
          title: Turrets L
        turrets_xl:
          type: integer
          title: Turrets Xl
        shields_s:
          type: integer
          title: Shields S
        shields_m:
          type: integer
          title: Shields M
        shields_l:
          type: integer
          title: Shields L
        shields_xl:
          type: integer
          title: Shields Xl
      additionalProperties: false
      type: object
      required:
      - module_id
      - name
      - dlc
      - kind
      - size
      - produces_ware_id
      - storage_capacity
      - storage_type
      - drone_capacity
      - workforce_capacity
      - icon_url
      - hull
      - explosiondamage
      - turrets_s
      - turrets_m
      - turrets_l
      - turrets_xl
      - shields_s
      - shields_m
      - shields_l
      - shields_xl
      title: ModuleDetail
    ModuleSummary:
      properties:
        module_id:
          type: string
          title: Module Id
        name:
          type: string
          title: Name
        dlc:
          anyOf:
          - type: string
          - type: 'null'
          title: Dlc
        kind:
          anyOf:
          - type: string
          - type: 'null'
          title: Kind
        size:
          anyOf:
          - type: string
          - type: 'null'
          title: Size
        produces_ware_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Produces Ware Id
        storage_capacity:
          anyOf:
          - type: integer
          - type: 'null'
          title: Storage Capacity
        storage_type:
          anyOf:
          - type: string
          - type: 'null'
          title: Storage Type
        drone_capacity:
          anyOf:
          - type: integer
          - type: 'null'
          title: Drone Capacity
        workforce_capacity:
          anyOf:
          - type: integer
          - type: 'null'
          title: Workforce Capacity
        icon_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Icon Url
      additionalProperties: false
      type: object
      required:
      - module_id
      - name
      - dlc
      - kind
      - size
      - produces_ware_id
      - storage_capacity
      - storage_type
      - drone_capacity
      - workforce_capacity
      - icon_url
      title: ModuleSummary
    NpcStationDetail:
      properties:
        station_id:
          type: string
          title: Station Id
        owner_faction:
          anyOf:
          - type: string
          - type: 'null'
          title: Owner Faction
        race:
          anyOf:
          - type: string
          - type: 'null'
          title: Race
        tags:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Tags
        location_sector:
          anyOf:
          - type: string
          - type: 'null'
          title: Location Sector
        location_zone:
          anyOf:
          - type: string
          - type: 'null'
          title: Location Zone
        x:
          anyOf:
          - type: number
          - type: 'null'
          title: X
        y:
          anyOf:
          - type: number
          - type: 'null'
          title: Y
        z:
          anyOf:
          - type: number
          - type: 'null'
          title: Z
      additionalProperties: false
      type: object
      required:
      - station_id
      - owner_faction
      - race
      - tags
      - location_sector
      - location_zone
      - x
      - y
      - z
      title: NpcStationDetail
    NpcStationSummary:
      properties:
        station_id:
          type: string
          title: Station Id
        owner_faction:
          anyOf:
          - type: string
          - type: 'null'
          title: Owner Faction
        race:
          anyOf:
          - type: string
          - type: 'null'
          title: Race
        tags:
          anyOf:
          - items:
              type: string
            type: array
          - type: 'null'
          title: Tags
        location_sector:
          anyOf:
          - type: string
          - type: 'null'
          title: Location Sector
      additionalProperties: false
      type: object
      required:
      - station_id
      - owner_faction
      - race
      - tags
      - location_sector
      title: NpcStationSummary
...
    ShipDetail:
      properties:
        ship_id:
          type: string
          title: Ship Id
        name:
          type: string
          title: Name
        dlc:
          anyOf:
          - type: string
          - type: 'null'
          title: Dlc
        class_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Class Id
        ship_type:
          anyOf:
          - type: string
          - type: 'null'
          title: Ship Type
        faction_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Faction Id
        role:
          anyOf:
          - type: string
          - type: 'null'
          title: Role
        hull:
          anyOf:
          - type: integer
          - type: 'null'
          title: Hull
        cargo_volume:
          anyOf:
          - type: integer
          - type: 'null'
          title: Cargo Volume
        speed_max:
          anyOf:
          - type: number
          - type: 'null'
          title: Speed Max
        icon_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Icon Url
        image_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Image Url
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        basename:
          anyOf:
          - type: string
          - type: 'null'
          title: Basename
        secrecy_level:
          anyOf:
          - type: integer
          - type: 'null'
          title: Secrecy Level
        mass:
          anyOf:
          - type: number
          - type: 'null'
          title: Mass
        drag_forward:
          anyOf:
          - type: number
          - type: 'null'
          title: Drag Forward
        drag_reverse:
          anyOf:
          - type: number
          - type: 'null'
          title: Drag Reverse
        drag_horizontal:
          anyOf:
          - type: number
          - type: 'null'
          title: Drag Horizontal
        drag_vertical:
          anyOf:
          - type: number
          - type: 'null'
          title: Drag Vertical
        drag_pitch:
          anyOf:
          - type: number
          - type: 'null'
          title: Drag Pitch
        drag_yaw:
          anyOf:
          - type: number
          - type: 'null'
          title: Drag Yaw
        drag_roll:
          anyOf:
          - type: number
          - type: 'null'
          title: Drag Roll
        inertia_pitch:
          anyOf:
          - type: number
          - type: 'null'
          title: Inertia Pitch
        inertia_yaw:
          anyOf:
          - type: number
          - type: 'null'
          title: Inertia Yaw
        inertia_roll:
          anyOf:
          - type: number
          - type: 'null'
          title: Inertia Roll
        people_capacity:
          anyOf:
          - type: integer
          - type: 'null'
          title: People Capacity
        missile_storage:
          anyOf:
          - type: integer
          - type: 'null'
          title: Missile Storage
        drone_storage:
          anyOf:
          - type: integer
          - type: 'null'
          title: Drone Storage
        countermeasure_storage:
          anyOf:
          - type: integer
          - type: 'null'
          title: Countermeasure Storage
        deployable_storage:
          anyOf:
          - type: integer
          - type: 'null'
          title: Deployable Storage
        weapons_s:
          type: integer
          title: Weapons S
        weapons_m:
          type: integer
          title: Weapons M
        weapons_l:
          type: integer
          title: Weapons L
        weapons_xl:
          type: integer
          title: Weapons Xl
        turrets_s:
          type: integer
          title: Turrets S
        turrets_m:
          type: integer
          title: Turrets M
        turrets_l:
          type: integer
          title: Turrets L
        turrets_xl:
          type: integer
          title: Turrets Xl
        shields_s:
          type: integer
          title: Shields S
        shields_m:
          type: integer
          title: Shields M
        shields_l:
          type: integer
          title: Shields L
        shields_xl:
          type: integer
          title: Shields Xl
        engines_s:
          type: integer
          title: Engines S
        engines_m:
          type: integer
          title: Engines M
        engines_l:
          type: integer
          title: Engines L
        engines_xl:
          type: integer
          title: Engines Xl
        software:
          items:
            $ref: '#/components/schemas/ShipSoftware'
          type: array
          title: Software
        drop_list_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Drop List Id
      additionalProperties: false
      type: object
      required:
      - ship_id
      - name
      - dlc
      - class_id
      - ship_type
      - faction_id
      - role
      - hull
      - cargo_volume
      - speed_max
      - icon_url
      - image_url
      - description
      - basename
      - secrecy_level
      - mass
      - drag_forward
      - drag_reverse
      - drag_horizontal
      - drag_vertical
      - drag_pitch
      - drag_yaw
      - drag_roll
      - inertia_pitch
      - inertia_yaw
      - inertia_roll
      - people_capacity
      - missile_storage
      - drone_storage
      - countermeasure_storage
      - deployable_storage
      - weapons_s
      - weapons_m
      - weapons_l
      - weapons_xl
      - turrets_s
      - turrets_m
      - turrets_l
      - turrets_xl
      - shields_s
      - shields_m
      - shields_l
      - shields_xl
      - engines_s
      - engines_m
      - engines_l
      - engines_xl
      - software
      - drop_list_id
      title: ShipDetail
    ShipSoftware:
      properties:
        ware_id:
          type: string
          title: Ware Id
        compatible:
          type: integer
          title: Compatible
        is_default:
          type: integer
          title: Is Default
      additionalProperties: false
      type: object
      required:
      - ware_id
      - compatible
      - is_default
      title: ShipSoftware
    ShipSummary:
      properties:
        ship_id:
          type: string
          title: Ship Id
        name:
          type: string
          title: Name
        dlc:
          anyOf:
          - type: string
          - type: 'null'
          title: Dlc
        class_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Class Id
        ship_type:
          anyOf:
          - type: string
          - type: 'null'
          title: Ship Type
        faction_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Faction Id
        role:
          anyOf:
          - type: string
          - type: 'null'
          title: Role
        hull:
          anyOf:
          - type: integer
          - type: 'null'
          title: Hull
        cargo_volume:
          anyOf:
          - type: integer
          - type: 'null'
          title: Cargo Volume
        speed_max:
          anyOf:
          - type: number
          - type: 'null'
          title: Speed Max
        icon_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Icon Url
        image_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Image Url
      additionalProperties: false
      type: object
      required:
      - ship_id
      - name
      - dlc
      - class_id
      - ship_type
      - faction_id
      - role
      - hull
      - cargo_volume
      - speed_max
      - icon_url
      - image_url
      title: ShipSummary
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...